삭제부분 수정함

+++ src/main/java/org/example/WordCRUD.java
@@ -99,7 +99,7 @@ public class WordCRUD implements ICRUD{
         System.out.print("정말로 삭제하시겠습니까?(Y/n) ");
         String answer = s.next();
         if(answer.equalsIgnoreCase("y")) {
-            list.remove(idlist.get(id - 1));
+            list.remove((int)idlist.get(id - 1));
             System.out.println("단어가 삭제되었습니다.");
         }else{
             System.out.println("취소되었습니다. ");
